refact: move authz code to common place
