+++ chord_metadata_service/metadata/settings.py
@@ -156,7 +156,7 @@ MIDDLEWARE = [
     'django.middleware.common.CommonMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'chord_metadata_service.metadata.authz.AuthzMiddleware'
+    'chord_metadata_service.authz.middleware.AuthzMiddleware'
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
